requirements.txt
yt-dlp
requests
anthropic
stripe
playwright
pandas
beautifulsoup4
scrapy
python-dotenv
google-generativeai
pyairtable
